requirements.txt
Babel==2.12.1
cobble==0.1.3
docxcompose==1.4.0
docxtpl==0.16.6
et-xmlfile==1.1.0
guidata==2.3.1
guiqwt==4.3.3
Jinja2==3.1.2
lxml==4.9.2
mammoth==1.5.1
MarkupSafe==2.1.2
numpy==1.24.2
openpyxl==3.1.2
packaging==23.1
Pillow==9.5.0
pyqt-tools==1.0.0
PyQt5==5.15.9
PyQt5-Qt5==5.15.2
PyQt5-sip==12.12.1
PyQt6==6.5.0
PyQt6-Qt6==6.5.0
PyQt6-sip==13.5.1
python-docx==0.8.11
PythonQwt==0.10.2
QtPy==2.3.1
scipy==1.10.1
six==1.16.0
